python_core/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
litellm>=1.74.0
python-dotenv==1.0.1
Pillow==10.4.0
requests==2.32.3
gunicorn==22.0.0 